Removing script browserify task, replacing it with the webpack test. Hooking up the gulp watch task to call webpacks watch feature.
