+++ package.json
@@ -38,12 +38,8 @@
     "jshint-stylish": "2.0.1",
     "gulp-imagemin": "2.3.0",
     "gulp-webserver": "0.9.1",
-    "vinyl-buffer": "1.0.0",
-    "vinyl-source-stream": "1.1.0",
     "gulp-sass": "2.0.3",
     "gulp-autoprefixer": "2.3.1",
-    "browserify": "10.2.6",
-    "gulp-uglify": "1.2.0",
     "webpack-stream": "2.1.1",
     "babel-loader": "6.0.1",
     "babel-core": "6.1.2",
